# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/sessions/{sessionId}/attendance-candidates": {
      "get": {
        "operationId": "AttendancesController_candidates",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceCandidates"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the trainees you can still add to this session, with the spots left.",
        "tags": [
          "Attendances"
        ]
      }
    },
    "/api/sessions/{sessionId}/attendances": {
      "get": {
        "operationId": "AttendancesController_list",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttendanceWithTraineeList"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the attendance rows of one session.",
        "tags": [
          "Attendances"
        ]
      },
      "post": {
        "operationId": "AttendancesController_addTrainee",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAttendanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attendance"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Add a trainee to a session. A prepaid visit is drawn if a card applies.",
        "tags": [
          "Attendances"
        ]
      },
      "put": {
        "operationId": "AttendancesController_bulkMark",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkMarkAttendancesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkMarkResult"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark the attendance of a whole session in one request.",
        "tags": [
          "Attendances"
        ]
      }
    },
    "/api/sessions/{sessionId}/attendances/{id}": {
      "delete": {
        "operationId": "AttendancesController_remove",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a booking. The visit returns and the queue head can get the spot.",
        "tags": [
          "Attendances"
        ]
      }
    },
    "/api/sessions/{sessionId}/rsvp": {
      "patch": {
        "operationId": "AttendancesController_rsvp",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RsvpDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attendance"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Record the reply of a customer for one session.",
        "tags": [
          "Attendances"
        ]
      }
    },
    "/api/me/sessions": {
      "get": {
        "operationId": "AttendancesController_myUpcoming",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSessionEntryList"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List the sessions of the family, with bookings and queue positions.",
        "tags": [
          "Attendances"
        ]
      }
    },
    "/api/me/sessions/{sessionId}/bookings": {
      "post": {
        "operationId": "AttendancesController_bookForCustomer",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAttendanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Attendance"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Book a spot as a customer. Needs self-booking on the class.",
        "tags": [
          "Attendances"
        ]
      }
    },
    "/api/me/sessions/{sessionId}/bookings/{traineeId}": {
      "delete": {
        "operationId": "AttendancesController_cancelForCustomer",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "traineeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "required": false,
            "description": "Tenant the caller acts in. A tenant-bound user must name one of their own tenants (403 otherwise); a SUPER_ADMIN may name any active tenant (404 otherwise).",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Cancel a booking as a customer. The visit returns.",
        "tags": [
          "Attendances"
        ]
      }
    },
...
  },
  "info": {
    "title": "PulseDesk API",
...
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AddAttendanceDto": {
        "type": "object",
        "properties": {
          "traineeId": {
            "type": "string"
          }
        },
        "required": [
          "traineeId"
        ]
      },
      "BulkMarkAttendanceItemDto": {
        "type": "object",
        "properties": {
          "traineeId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PRESENT",
              "ABSENT",
              "EXCUSED"
            ]
          },
          "notes": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "required": [
          "traineeId",
          "status"
        ]
      },
      "BulkMarkAttendancesDto": {
        "type": "object",
        "properties": {
          "items": {
            "uniqueItems": true,
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BulkMarkAttendanceItemDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
      "RsvpDto": {
        "type": "object",
        "properties": {
          "traineeId": {
            "type": "string"
          },
          "traineeRsvp": {
            "type": "string",
            "enum": [
              "CONFIRMED",
              "DECLINED",
              "RESCHEDULE_REQUESTED"
            ]
          }
        },
        "required": [
          "traineeId",
          "traineeRsvp"
        ]
      },
...
      "AttendanceCandidates": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "tenantId": {
                  "type": "string"
                },
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                },
                "dateOfBirth": {
                  "type": "string"
                },
                "phone": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "email": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "notes": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "isActive": {
                  "type": "boolean"
                },
                "userId": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "createdAt": {
                  "type": "string"
                },
                "updatedAt": {
                  "type": "string"
                },
                "card": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "visitsRemaining": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "expiresAt": {
                          "anyOf": [
                            {
                              "type": "null"
                            },
                            {
                              "type": "string"
                            }
                          ]
                        },
                        "classScoped": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "id",
                        "visitsRemaining",
                        "expiresAt",
                        "classScoped"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "hasCards": {
                  "type": "boolean"
                }
              },
              "required": [
                "id",
                "tenantId",
                "firstName",
                "lastName",
                "dateOfBirth",
                "phone",
                "email",
                "notes",
                "isActive",
                "userId",
                "createdAt",
                "updatedAt",
                "card",
                "hasCards"
              ],
              "additionalProperties": false
            }
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "total": {
            "type": "number"
          },
          "totalPages": {
            "type": "number"
          },
          "spotsLeft": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "totalPages",
          "spotsLeft"
        ],
        "additionalProperties": false
      },
      "AttendanceWithTraineeList": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "tenantId": {
              "type": "string"
            },
            "sessionId": {
              "type": "string"
            },
            "traineeId": {
              "type": "string"
            },
            "status": {
              "type": "string",
              "enum": [
                "PENDING",
                "PRESENT",
                "ABSENT",
                "EXCUSED"
              ]
            },
            "traineeRsvp": {
              "anyOf": [
                {
                  "type": "string",
                  "enum": [
                    "CONFIRMED",
                    "DECLINED",
                    "RESCHEDULE_REQUESTED"
                  ]
                },
                {
                  "type": "null"
                }
              ]
            },
            "notes": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "markedAt": {
              "anyOf": [
                {
                  "type": "null"
                },
                {
                  "type": "string"
                }
              ]
            },
            "markedById": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "markedByEmailSnapshot": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "markedByNameSnapshot": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "createdAt": {
              "type": "string"
            },
            "updatedAt": {
              "type": "string"
            },
            "trainee": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "firstName",
                "lastName"
              ],
              "additionalProperties": false
            }
          },
          "required": [
            "id",
            "tenantId",
            "sessionId",
            "traineeId",
            "status",
            "traineeRsvp",
            "notes",
            "markedAt",
            "markedById",
            "markedByEmailSnapshot",
            "markedByNameSnapshot",
            "createdAt",
            "updatedAt",
            "trainee"
          ],
          "additionalProperties": false
        }
      },
      "Attendance": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "traineeId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "PRESENT",
              "ABSENT",
              "EXCUSED"
            ]
          },
          "traineeRsvp": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "CONFIRMED",
                  "DECLINED",
                  "RESCHEDULE_REQUESTED"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "markedAt": {
            "anyOf": [
              {
                "type": "null"
              },
              {
                "type": "string"
              }
            ]
          },
          "markedById": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "markedByEmailSnapshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "markedByNameSnapshot": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "sessionId",
          "traineeId",
          "status",
          "traineeRsvp",
          "notes",
          "markedAt",
          "markedById",
          "markedByEmailSnapshot",
          "markedByNameSnapshot",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "BulkMarkResult": {
        "type": "object",
        "properties": {
          "updated": {
            "type": "number"
          }
        },
        "required": [
          "updated"
        ],
        "additionalProperties": false
      },
...
    }
  }
}